Split AppFeature reducer handlers
