+++ supacode/Features/App/Reducer/AppFeature.swift
@@ -4,43 +4,6 @@ import Foundation
 import PostHog
 import SwiftUI
 
-private let appLogger = SupaLogger("App")
-private let notificationJumpLogger = SupaLogger("NotificationJump")
-
-private enum CancelID {
-  static let periodicRefresh = "app.periodicRefresh"
-}
-
-private func makeTerminalRestorableWorktrees(from repositories: [Repository]) -> [Worktree] {
-  var worktrees: [Worktree] = []
-  worktrees.reserveCapacity(repositories.reduce(0) { $0 + max(1, $1.worktrees.count) })
-  for repository in repositories {
-    if repository.capabilities.supportsWorktrees {
-      worktrees.append(contentsOf: repository.worktrees)
-      continue
-    }
-    if repository.capabilities.supportsRunnableFolderActions {
-      worktrees.append(
-        Worktree(
-          id: repository.id,
-          name: repository.name,
-          detail: repository.rootURL.path(percentEncoded: false),
-          workingDirectory: repository.rootURL,
-          repositoryRootURL: repository.rootURL
-        )
-      )
-    }
-  }
-  return worktrees
-}
-
-private func openedWorktreeIDsForInfoWatcher(
-  from repositories: RepositoriesFeature.State
-) -> Set<Worktree.ID> {
-  let watchedIDs = Set(repositories.worktreesForInfoWatcher().map(\.id))
-  return repositories.openedWorktreeIDs.intersection(watchedIDs)
-}
-
 @Reducer
 struct AppFeature {
   @ObservableState
@@ -122,224 +85,18 @@ struct AppFeature {
     case confirmQuit
   }
 
-  @Dependency(AnalyticsClient.self) private var analyticsClient
-  @Dependency(\.date.now) private var now
-  @Dependency(RepositoryPersistenceClient.self) private var repositoryPersistence
-  @Dependency(WorkspaceClient.self) private var workspaceClient
-  @Dependency(SettingsWindowClient.self) private var settingsWindowClient
-  @Dependency(AppLifecycleClient.self) private var appLifecycleClient
-  @Dependency(NotificationSoundClient.self) private var notificationSoundClient
-  @Dependency(SystemNotificationClient.self) private var systemNotificationClient
-  @Dependency(DockClient.self) private var dockClient
-  @Dependency(TerminalClient.self) private var terminalClient
-  @Dependency(WorktreeInfoWatcherClient.self) private var worktreeInfoWatcher
-  @Dependency(CustomShortcutRegistryClient.self) private var customShortcutRegistryClient
-
-  private func appQuitProperties(launchedAt: Date?) -> [String: Any]? {
-    guard let seconds = Self.sessionDurationSeconds(launchedAt: launchedAt, now: now) else { return nil }
-    return ["session_duration_seconds": seconds]
-  }
-
-  private func restoreCommandPaletteTerminalFocusEffect(repositories: RepositoriesFeature.State) -> Effect<Action> {
-    guard let worktree = commandPaletteTerminalFocusTarget(repositories: repositories) else { return .none }
-    return .run { _ in
-      await terminalClient.send(.focusSelectedTab(worktree))
-    }
-  }
-
-  /// Delegate actions that intentionally shift focus elsewhere (selection
-  /// changes, view changes that swap the focused terminal). These skip the
-  /// post-delegate focus restore so we don't fight the natural new focus.
-  private static func commandPaletteDelegateChangesActiveSelection(
-    _ delegate: CommandPaletteFeature.Delegate
-  ) -> Bool {
-    switch delegate {
-    case .selectWorktree, .jumpToLatestUnread, .viewArchivedWorktrees,
-      .newWorktree, .toggleCanvas, .renameBranch:
-      return true
-    default:
-      return false
-    }
-  }
-
-  private func commandPaletteTerminalFocusTarget(repositories: RepositoriesFeature.State) -> Worktree? {
-    if let worktree = repositories.selectedTerminalWorktree {
-      return worktree
-    }
-    return canvasFocusedTerminalWorktree(repositories: repositories)
-  }
-
-  private func actionTargetWorktree(repositories: RepositoriesFeature.State) -> Worktree? {
-    if let worktree = repositories.selectedTerminalWorktree {
-      return worktree
-    }
-    return canvasFocusedTerminalWorktree(repositories: repositories)
-  }
-
-  private func canvasFocusedTerminalWorktree(repositories: RepositoriesFeature.State) -> Worktree? {
-    guard repositories.isShowingCanvas,
-      let worktreeID = terminalClient.canvasFocusedWorktreeID()
-    else {
-      return nil
-    }
-    return terminalWorktree(for: worktreeID, repositories: repositories)
-  }
-
-  private func terminalWorktree(
-    for worktreeID: Worktree.ID,
-    repositories: RepositoriesFeature.State
-  ) -> Worktree? {
-    if let worktree = repositories.worktree(for: worktreeID) {
-      return worktree
-    }
-    guard let repository = repositories.repositories[id: worktreeID],
-      repository.capabilities.supportsRunnableFolderActions,
-      !repository.capabilities.supportsWorktrees
-    else {
-      return nil
-    }
-    return Worktree(
-      id: repository.id,
-      name: repository.name,
-      detail: repository.rootURL.path(percentEncoded: false),
-      workingDirectory: repository.rootURL,
-      repositoryRootURL: repository.rootURL
-    )
-  }
-
-  static func sessionDurationSeconds(launchedAt: Date?, now: Date) -> Int? {
-    guard let launchedAt else { return nil }
-    return max(0, Int(now.timeIntervalSince(launchedAt)))
-  }
-
-  private func resolvedKeybindings(
-    settings: SettingsFeature.State,
-    customCommands: [UserCustomCommand]
-  ) -> ResolvedKeybindingMap {
-    let migration = LegacyCustomCommandShortcutMigration.migrate(commands: customCommands)
-    var resolved = KeybindingResolver.resolve(
-      schema: .appResolverSchema(customCommands: customCommands),
-      userOverrides: settings.keybindingUserOverrides,
-      migratedOverrides: migration.overrides
-    )
-    let customCommandIDs = customCommands.map { command in
-      LegacyCustomCommandShortcutMigration.customCommandBindingID(for: command.id)
-    }
-    let customCommandBindings = customCommandIDs.compactMap { resolved.keybinding(for: $0) }
-    guard !customCommandBindings.isEmpty else {
-      return resolved
-    }
-    for binding in AppShortcuts.bindings where binding.scope == .configurableAppAction {
-      guard let resolvedBinding = resolved.binding(for: binding.id),
-        let shortcut = resolvedBinding.binding,
-        customCommandBindings.contains(shortcut)
-      else {
-        continue
-      }
-      resolved.bindingsByCommandID[binding.id] = ResolvedKeybinding(
-        command: resolvedBinding.command,
-        binding: nil,
-        source: resolvedBinding.source
-      )
-    }
-    return resolved
-  }
-
-  /// Applies a worktree's repository settings (open action, run script) into
-  /// state. Shared by the normal `worktreeSettingsLoaded` action and the Canvas
-  /// focus path so both stay in sync.
-  private func applyWorktreeSettings(_ settings: RepositorySettings, into state: inout State) {
-    @Shared(.settingsFile) var settingsFile
-    let normalizedDefaultEditorID = OpenWorktreeAction.normalizedDefaultEditorID(
-      settingsFile.global.defaultEditorID
-    )
-    state.openActionSelection = OpenWorktreeAction.fromSettingsID(
-      settings.openActionID,
-      defaultEditorID: normalizedDefaultEditorID
-    )
-    state.selectedRunScript = settings.runScript
-  }
-
-  /// Applies a worktree's user settings (custom commands, keybindings) into
-  /// state and returns the effect that re-registers custom shortcuts. Shared by
-  /// the normal `worktreeUserSettingsLoaded` action and the Canvas focus path.
-  private func applyWorktreeUserSettings(
-    _ settings: UserRepositorySettings,
-    into state: inout State
-  ) -> Effect<Action> {
-    state.selectedCustomCommands = UserRepositorySettings.normalizedCommands(settings.customCommands)
-    state.resolvedKeybindings = resolvedKeybindings(
-      settings: state.settings,
-      customCommands: state.selectedCustomCommands
-    )
-    let userOverrideConflicts = AppShortcuts.userOverrideConflicts(in: state.selectedCustomCommands)
-    let shortcuts: [UserCustomShortcut] = state.selectedCustomCommands.compactMap { command in
-      let commandID = LegacyCustomCommandShortcutMigration.customCommandBindingID(for: command.id)
-      return state.resolvedKeybindings.keybinding(for: commandID)?.userCustomShortcut
-    }
-    return .run { _ in
-      let logger = SupaLogger("Shortcuts")
-      for conflict in userOverrideConflicts {
-        logger.warning(
-          "shortcut_conflict reason=userOverride app_action=\"\(conflict.appActionTitle)\" "
-            + "app_shortcut=\(conflict.appShortcutDisplay) custom_command=\"\(conflict.commandTitle)\" "
-            + "custom_shortcut=\(conflict.commandShortcutDisplay) result=customOverride"
-        )
-      }
-      await customShortcutRegistryClient.setShortcuts(shortcuts)
-    }
-  }
-
-  private func applyDefaultViewMode(into state: inout State) -> Effect<Action> {
-    guard !state.hasAppliedInitialViewMode else { return .none }
-    state.hasAppliedInitialViewMode = true
-
-    @Shared(.settingsFile) var settingsFile
-    let shouldEnterShelf =
-      settingsFile.global.defaultViewMode == .shelf
-      && !state.repositories.isShelfActive
-    let shouldEnterCanvas =
-      settingsFile.global.defaultViewMode == .canvas
-      && !state.repositories.isShowingCanvas
-
-    var effects: [Effect<Action>] = []
-    if shouldEnterShelf {
-      effects.append(.send(.repositories(.toggleShelf)))
-    }
-    // Enter Canvas after the selection effects so `.selectCanvas`
-    // records the just-selected worktree as the pre-Canvas anchor.
-    if shouldEnterCanvas {
-      if state.repositories.selectedTerminalWorktree == nil,
-        let targetID = defaultViewFallbackSelectionID(in: state.repositories)
-      {
-        effects.append(defaultViewSelectionEffect(for: targetID, repositories: state.repositories))
-      }
-      effects.append(.send(.repositories(.toggleCanvas)))
-    }
-    return effects.isEmpty ? .none : .concatenate(effects)
-  }
-
-  private func defaultViewFallbackSelectionID(in repositories: RepositoriesFeature.State) -> Worktree.ID? {
-    let candidateIDs =
-      [repositories.lastFocusedWorktreeID].compactMap(\.self)
-      + repositories.orderedWorktreeRows().map(\.id)
-    return candidateIDs.first { id in
-      repositories.worktree(for: id) != nil
-        || repositories.repositories[id: id]?.kind == .plain
-    }
-  }
-
-  private func defaultViewSelectionEffect(
-    for targetID: Worktree.ID,
-    repositories: RepositoriesFeature.State
-  ) -> Effect<Action> {
-    if repositories.worktree(for: targetID) == nil,
-      repositories.repositories[id: targetID]?.kind == .plain
-    {
-      return .send(.repositories(.selectRepository(targetID)))
-    }
-    return .send(.repositories(.selectWorktree(targetID)))
-  }
+  @Dependency(AnalyticsClient.self) var analyticsClient
+  @Dependency(\.date.now) var now
+  @Dependency(RepositoryPersistenceClient.self) var repositoryPersistence
+  @Dependency(WorkspaceClient.self) var workspaceClient
+  @Dependency(SettingsWindowClient.self) var settingsWindowClient
+  @Dependency(AppLifecycleClient.self) var appLifecycleClient
+  @Dependency(NotificationSoundClient.self) var notificationSoundClient
+  @Dependency(SystemNotificationClient.self) var systemNotificationClient
+  @Dependency(DockClient.self) var dockClient
+  @Dependency(TerminalClient.self) var terminalClient
+  @Dependency(WorktreeInfoWatcherClient.self) var worktreeInfoWatcher
+  @Dependency(CustomShortcutRegistryClient.self) var customShortcutRegistryClient
 
   var body: some Reducer<State, Action> {
     let core = Reduce<State, Action> { state, action in
@@ -1147,347 +904,11 @@ struct AppFeature {
       case .updates:
         return .none
 
-      case .commandPalette(.setPresented(false)):
-        guard state.commandPalette.isPresented else { return .none }
-        return restoreCommandPaletteTerminalFocusEffect(repositories: state.repositories)
+      case .commandPalette(let action):
+        return reduceCommandPaletteAction(action, state: &state)
 
-      case .commandPalette(.togglePresented):
-        guard state.commandPalette.isPresented else { return .none }
-        return restoreCommandPaletteTerminalFocusEffect(repositories: state.repositories)
-
-      case .commandPalette(.delegate(.selectWorktree(let worktreeID))):
-        if state.repositories.isShowingCanvas {
-          if state.repositories.worktree(for: worktreeID) == nil,
-            state.repositories.repositories[id: worktreeID]?.kind == .plain
-          {
-            return .send(.repositories(.focusCanvasRepository(worktreeID)))
-          }
-          return .send(.repositories(.focusCanvasWorktree(worktreeID)))
-        }
-        return .send(.repositories(.selectWorktree(worktreeID)))
-
-      case .commandPalette(.delegate(.checkForUpdates)):
-        return .send(.updates(.checkForUpdates))
-
-      case .commandPalette(.delegate(.openSettings)):
-        return .merge(
-          .send(.settings(.setSelection(.general))),
-          .run { _ in
-            await settingsWindowClient.show()
-          }
-        )
-
-      case .commandPalette(.delegate(.newWorktree)):
-        return .send(.repositories(.worktreeCreation(.createRandomWorktree)))
-
-      case .commandPalette(.delegate(.openRepository)):
-        return .send(.repositories(.setOpenPanelPresented(true)))
-
-      case .commandPalette(.delegate(.deleteWorktree(let worktreeID, let repositoryID))):
-        return .send(.repositories(.worktreeLifecycle(.requestDeleteWorktree(worktreeID, repositoryID))))
-
-      case .commandPalette(.delegate(.viewArchivedWorktrees)):
-        return .send(.repositories(.selectArchivedWorktrees))
-
-      case .commandPalette(.delegate(.refreshWorktrees)):
-        return .send(.repositories(.refreshWorktrees))
-
-      case .commandPalette(.delegate(.jumpToLatestUnread)):
-        return .send(.jumpToLatestUnread)
-
-      case .commandPalette(.delegate(.installCLI)):
-        return .send(.settings(.installCLIButtonTapped(showAlert: false)))
-
-      case .commandPalette(.delegate(.toggleLeftSidebar)):
-        return .send(.toggleLeftSidebar)
-
-      case .commandPalette(.delegate(.toggleActiveAgentsPanel)):
-        return .send(.repositories(.activeAgents(.togglePanelVisibility)))
-
-      case .commandPalette(.delegate(.toggleCanvas)):
-        return .send(.repositories(.toggleCanvas))
-
-      case .commandPalette(.delegate(.expandCanvasCard)):
-        return .send(.repositories(.requestCanvasCommand(.toggleExpand)))
-
-      case .commandPalette(.delegate(.arrangeCanvasCards)):
-        return .send(.repositories(.requestCanvasCommand(.arrange)))
-
-      case .commandPalette(.delegate(.organizeCanvasCards)):
-        return .send(.repositories(.requestCanvasCommand(.organize)))
-
-      case .commandPalette(.delegate(.selectAllCanvasCards)):
-        return .send(.repositories(.requestCanvasCommand(.selectAll)))
-
-      case .commandPalette(.delegate(.toggleShelf)):
-        return .send(.repositories(.toggleShelf))
-
-      case .commandPalette(.delegate(.showDiff)):
-        guard let worktreeID = state.repositories.selectedWorktreeID,
-          let worktree = state.repositories.worktree(for: worktreeID)
-        else {
-          return .none
-        }
-        let keybindings = state.resolvedKeybindings
-        return .run { _ in
-          await MainActor.run {
-            DiffWindowManager.shared.show(
-              worktreeURL: worktree.workingDirectory,
-              branchName: worktree.name,
-              resolvedKeybindings: keybindings
-            )
-          }
-        }
-
-      case .commandPalette(.delegate(.revealInFinder)):
-        return .send(.openWorktree(.finder))
-
-      case .commandPalette(.delegate(.copyPath)):
-        guard let worktree = state.repositories.selectedTerminalWorktree else {
-          return .none
-        }
-        let path = worktree.workingDirectory.path
-        return .run { _ in
-          await MainActor.run {
-            NSPasteboard.general.clearContents()
-            NSPasteboard.general.setString(path, forType: .string)
-          }
-        }
-
-      case .commandPalette(.delegate(.revealInSidebar)):
-        guard state.repositories.selectedWorktreeID != nil else { return .none }
-        return .merge(
-          .send(.showLeftSidebar),
-          .send(.repositories(.revealSelectedWorktreeInSidebar))
-        )
-
-      case .commandPalette(.delegate(.runScript)):
-        return .send(.runScript)
-
-      case .commandPalette(.delegate(.stopRunScript)):
-        return .send(.stopRunScript)
-
-      case .commandPalette(.delegate(.renameBranch)):
-        guard let worktreeID = state.repositories.selectedWorktreeID else { return .none }
-        return .send(.repositories(.requestRenameBranchPrompt(worktreeID)))
-
-      case .commandPalette(.delegate(.openRepositorySettings(let repositoryID))):
-        // Reuse the existing repo-side flow so the repo-existence guard and
-        // settingsWindowClient.show() live in one place
-        // (.repositories(.delegate(.openRepositorySettings(_))) at line ~441).
-        return .send(.repositories(.repositoryManagement(.openRepositorySettings(repositoryID))))
-
-      case .commandPalette(.delegate(.togglePinWorktree(let worktreeID, let isCurrentlyPinned))):
-        if isCurrentlyPinned {
-          return .send(.repositories(.worktreeOrdering(.unpinWorktree(worktreeID))))
-        }
-        return .send(.repositories(.worktreeOrdering(.pinWorktree(worktreeID))))
-
-      case .commandPalette(.delegate(.runCustomCommand(let index))):
-        return .send(.runCustomCommand(index))
-
-      case .commandPalette(.delegate(.ghosttyCommand(let action))):
-        guard let worktree = actionTargetWorktree(repositories: state.repositories) else {
-          return .none
-        }
-        return .run { _ in
-          await terminalClient.send(.performBindingAction(worktree, action: action))
-        }
-
-      case .commandPalette(.delegate(.changeFocusedTabIcon(let worktreeID))):
-        guard let worktree = state.repositories.selectedTerminalWorktree,
-          worktree.id == worktreeID
-        else {
-          return .none
-        }
-        return .run { _ in
-          await terminalClient.send(.presentTabIconPicker(worktree))
-        }
-
-      case .commandPalette(.delegate(.openPullRequest(let worktreeID))):
-        return .send(.repositories(.githubIntegration(.pullRequestAction(worktreeID, .openOnCodeHost))))
-
-      case .commandPalette(.delegate(.markPullRequestReady(let worktreeID))):
-        return .send(.repositories(.githubIntegration(.pullRequestAction(worktreeID, .markReadyForReview))))
-
-      case .commandPalette(.delegate(.mergePullRequest(let worktreeID))):
-        return .send(.repositories(.githubIntegration(.pullRequestAction(worktreeID, .merge))))
-
-      case .commandPalette(.delegate(.closePullRequest(let worktreeID))):
-        return .send(.repositories(.githubIntegration(.pullRequestAction(worktreeID, .close))))
-
-      case .commandPalette(.delegate(.copyFailingJobURL(let worktreeID))):
-        return .send(.repositories(.githubIntegration(.pullRequestAction(worktreeID, .copyFailingJobURL))))
-
-      case .commandPalette(.delegate(.copyCiFailureLogs(let worktreeID))):
-        return .send(.repositories(.githubIntegration(.pullRequestAction(worktreeID, .copyCiFailureLogs))))
-
-      case .commandPalette(.delegate(.rerunFailedJobs(let worktreeID))):
-        return .send(.repositories(.githubIntegration(.pullRequestAction(worktreeID, .rerunFailedJobs))))
-
-      case .commandPalette(.delegate(.openFailingCheckDetails(let worktreeID))):
-        return .send(.repositories(.githubIntegration(.pullRequestAction(worktreeID, .openFailingCheckDetails))))
-
-      #if DEBUG
-        case .commandPalette(.delegate(.debugTestToast(let toast))):
-          return .send(.repositories(.showToast(toast)))
-
-        case .commandPalette(.delegate(.debugSimulateUpdateFound)):
-          return .send(.updates(.debugSimulateUpdateFound))
-
-        case .commandPalette(.delegate(.debugLightDockNotificationDot)):
-          return .run { _ in await dockClient.setNotificationBadge(1) }
-      #endif
-
-      case .commandPalette:
-        return .none
-
-      case .terminalEvent(.customCommandSucceeded(_, let name, let durationMs)):
-        let message = "\(name) succeeded in \(formatCustomCommandDuration(durationMs))"
-        return .send(.repositories(.showToast(.success(message))))
-
-      case .terminalEvent(.notificationReceived(let worktreeID, let surfaceID, let title, let body)):
-        var effects: [Effect<Action>] = [
-          .send(.repositories(.worktreeOrdering(.worktreeNotificationReceived(worktreeID))))
-        ]
-        if state.settings.systemNotificationsEnabled {
-          effects.append(
-            .run { _ in
-              await systemNotificationClient.send(title, body, worktreeID, surfaceID)
-            }
-          )
-        }
-        if state.settings.notificationSoundEnabled && !state.settings.systemNotificationsEnabled {
-          effects.append(
-            .run { _ in
-              await notificationSoundClient.play()
-            }
-          )
-        }
-        let bounceMode = state.settings.dockBounceMode
-        if bounceMode != .off {
-          effects.append(
-            .run { _ in
-              await dockClient.bounce(bounceMode)
-            }
-          )
-        }
-        return .merge(effects)
-
-      case .terminalEvent(.notificationIndicatorChanged(let count)):
-        state.notificationIndicatorCount = count
-        let badgeCount = state.settings.showNotificationDotOnDock ? count : 0
-        return .run { _ in
-          await dockClient.setNotificationBadge(badgeCount)
-        }
-
-      case .terminalEvent(.runScriptStatusChanged(let worktreeID, let isRunning)):
-        if isRunning {
-          state.runScriptStatusByWorktreeID[worktreeID] = true
-        } else {
-          state.runScriptStatusByWorktreeID.removeValue(forKey: worktreeID)
-        }
-        return .none
-
-      case .terminalEvent(.agentEntryChanged(let entry)):
-        return .send(
-          .repositories(
-            .activeAgents(
-              .agentEntryChanged(entry, autoShowPanel: state.settings.autoShowActiveAgentsPanel)
-            )
-          )
-        )
-
-      case .terminalEvent(.agentEntryRemoved(let id)):
-        return .send(.repositories(.activeAgents(.agentEntryRemoved(id))))
-
-      case .terminalEvent(.commandPaletteToggleRequested(let worktreeID)):
-        if state.commandPalette.isPresented {
-          return .send(.commandPalette(.setPresented(false)))
-        }
-        if state.repositories.worktree(for: worktreeID) != nil {
-          return .merge(
-            .send(.repositories(.selectWorktree(worktreeID))),
-            .send(.commandPalette(.setPresented(true)))
-          )
-        }
-        if state.repositories.repositories[id: worktreeID]?.kind == .plain {
-          return .merge(
-            .send(.repositories(.selectRepository(worktreeID))),
-            .send(.commandPalette(.setPresented(true)))
-          )
-        }
-        return .send(.commandPalette(.setPresented(true)))
-      case .terminalEvent(.setupScriptConsumed(let worktreeID)):
-        return .send(.repositories(.worktreeCreation(.consumeSetupScript(worktreeID))))
-
-      case .terminalEvent(.fontSizeChanged(let fontSize)):
-        return .send(.settings(.setTerminalFontSize(fontSize)))
-
-      case .terminalEvent(.layoutRestored(let selectedWorktreeID)):
-        appLogger.info("[LayoutRestore] layoutRestored: selectedWorktreeID=\(selectedWorktreeID ?? "nil")")
-        // Layout restore has settled: tabs are re-created, selection is set.
-        // Now apply the default view preference, which was deferred in
-        // `repositoriesChanged` (via `shouldDeferDefaultView`) to avoid
-        // stray spines and a selection flash.
-        var effects: [Effect<Action>] = []
-        if let selectedWorktreeID {
-          // Plain folders use .repository selection, not .worktree
-          if let repo = state.repositories.repositories[id: selectedWorktreeID],
-            repo.kind == .plain
-          {
-            effects.append(.send(.repositories(.selectRepository(selectedWorktreeID))))
-          } else {
-            effects.append(.send(.repositories(.selectWorktree(selectedWorktreeID))))
-          }
-        }
-        return .concatenate([.merge(effects), applyDefaultViewMode(into: &state)])
-
-      case .terminalEvent(.layoutRestoreFailed(let message)):
-        appLogger.warning("[LayoutRestore] layoutRestoreFailed: \(message)")
-        return .merge(
-          .send(.repositories(.showToast(.warning(message)))),
-          applyDefaultViewMode(into: &state)
-        )
-
-      case .terminalEvent(.tabCreated(let worktreeID)):
-        // Every tab creation (user +, CLI open, layout restore, …)
-        // marks its worktree as Shelf-visible. Layout restore in
-        // particular only calls `selectWorktree` for the one active
-        // worktree; other restored worktrees only surface here.
-        var openedWorktreeIDs = openedWorktreeIDsForInfoWatcher(from: state.repositories)
-        if state.repositories.worktree(for: worktreeID) != nil {
-          openedWorktreeIDs.insert(worktreeID)
-        }
-        let syncedOpenedWorktreeIDs = openedWorktreeIDs
-        return .merge(
-          .send(.repositories(.markWorktreeOpened(worktreeID))),
-          .run { _ in
-            await worktreeInfoWatcher.send(.setOpenedWorktreeIDs(syncedOpenedWorktreeIDs))
-          }
-        )
-
-      case .terminalEvent(.tabClosed(let worktreeID, let remainingTabs)):
-        // Closing the last tab retires the book from the Shelf. Other
-        // closes are routine and need no Reducer-side bookkeeping.
-        guard remainingTabs == 0 else { return .none }
-        var openedWorktreeIDs = openedWorktreeIDsForInfoWatcher(from: state.repositories)
-        openedWorktreeIDs.remove(worktreeID)
-        let syncedOpenedWorktreeIDs = openedWorktreeIDs
-        return .merge(
-          .send(.repositories(.markWorktreeClosed(worktreeID))),
-          .run { _ in
-            await worktreeInfoWatcher.send(.setOpenedWorktreeIDs(syncedOpenedWorktreeIDs))
-          }
-        )
-
-      case .terminalEvent(.focusChanged(_, let surfaceID)):
-        // Keep the Active Agents panel's keyboard-navigation anchor in sync with
-        // the surface that actually has focus, so ⌃⌥↑/↓ steps from the right place.
-        return .send(.repositories(.activeAgents(.focusedSurfaceChanged(surfaceID))))
-
-      case .terminalEvent:
-        return .none
+      case .terminalEvent(let event):
+        return reduceTerminalEvent(event, state: &state)
       }
     }
     core
@@ -1516,18 +937,4 @@ struct AppFeature {
       CommandPaletteFeature()
     }
   }
-}
-
-// Renders Custom Command run duration for status toasts.
-// Sub-second runs show ms; short runs show one decimal; long runs reuse the
-// whole-seconds formatter used by other command-finished notifications.
-func formatCustomCommandDuration(_ durationMs: Int) -> String {
-  if durationMs < 1_000 {
-    return "\(max(durationMs, 0))ms"
-  }
-  let seconds = Double(durationMs) / 1_000.0
-  if seconds < 10 {
-    return String(format: "%.1fs", seconds)
-  }
-  return WorktreeTerminalState.formatDuration(Int(seconds))
 }
